refactor(cli): one shared platform-command runner instead of nine copies
organizations.ts was the ninth copy of addPlatformOptions/runPlatformCommand/
PlatformOptions, and the copies had already drifted three ways: eval grew a
webOrigin in the execute context, projects an external abort signal, chat a
formatting variant. Extract the superset into cli/src/lib/platform-command.ts —
the context always carries webOrigin (a pure string transform of the base URL)
and the external signal is optional — and point all nine command files at it.
projects keeps its local PlatformOptions extension (the `project` selector
flag) and its Commander option-precedence helpers, which are genuinely its own.

575 CLI tests pass; tsc clean.
